Fix Invalid Code
The "Avoid Side Effects" section's "Bad" code block is actually invalid code, `splitIntoFirstAndLastName()` will throw a notice of `Undefined variable`. We need a `$global` here.

I'm all for competing against code that people may have but there is no point comparing code that doesn't work in the first place, it's out of the bounds of this document.